requirements.txt
openai
chromadb>=0.5.4
python-dotenv
